Move components into dedicated folders
