+++ src/components/ScreenBannerDetails/BannerDetailsDropTable.vue
