+++ src/components/ScreenBannerDetails/BannerDetailsScreen.vue
