+++ src/components/ScreenBannerDetails/BannerStylisedText.vue
