+++ src/components/ScreenBannerDetails/WishDetailsH2.vue
