+++ src/components/WishBanners/WishButton.vue
