+++ src/components/overlays/ItemDescriptionOverlay.vue
