+++ src/components/overlays/ItemObtainOverlay.vue
@@ -262,7 +262,7 @@ export default defineComponent({
 
 .main::before {
   content: "";
-  background: url("../assets/images/wish-reveal-background.jpg") center/cover
+  background: url("../../assets/images/wish-reveal-background.jpg") center/cover
       no-repeat fixed,
     #4a4a52;
   top: 0;
