+++ src/components/shared/CloseButton.vue
