+++ src/components/shared/GemCounter.vue
